requirements.txt
pandas
yfinance
fredapi
python-dotenv
jinja2
matplotlib